+++ src/main/java/visang/dataplatform/dataportal/model/dto/stdcheck/request/CreateStdTermDto.java
@@ -4,10 +4,10 @@ import lombok.Getter;
 
 @Getter
 public class CreateStdTermDto {
-    private Long term_idx;
+    private Integer term_idx;
     private String term_logical_nm;
     private String term_logical_desc;
     private String term_physical_nm;
     private String term_physical_desc;
-    private Long domain_idx;
+    private Integer domain_idx;
 }
