+++ src/main/java/visang/dataplatform/dataportal/model/dto/stdcheck/request/UpdateStdTermDto.java
@@ -6,10 +6,10 @@ import lombok.Setter;
 @Getter
 @Setter
 public class UpdateStdTermDto {
-    private Long term_idx;
+    private Integer term_idx;
     private String term_logical_nm;
     private String term_logical_desc;
     private String term_physical_nm;
     private String term_physical_desc;
-    private Long domain_idx;
+    private Integer domain_idx;
 }
