+++ src/main/java/visang/dataplatform/dataportal/model/dto/stdcheck/request/UpdateStdWordDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class UpdateStdWordDto {
-    private Long word_idx;
+    private Integer word_idx;
     private String word_logical_nm;
     private String word_physical_nm;
     private String word_logical_desc;
