+++ src/main/java/visang/dataplatform/dataportal/model/dto/stdcheck/response/SimpleStdWordDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class SimpleStdWordDto {
-    private Long word_idx;
+    private Integer word_idx;
     private String word_logical_nm;
     private String word_physical_nm;
 }
